employee task review comments fixed. Небольшое замечания: Задания сформулированые не точно!
например: Емнип, менеджер есть только у работника.
В задании: У сотрудника (Employee) есть, кроме ФИО, ссылка на менеджера, которому он отчитывается. Заметьте, что у менеджера нет своего менеджера.

Задание 5:
почему task1?
ответ: Согласно "https://github.com/gladorange/mera_se_2018/tree/master/_ATARASOV"
"... на каждое задание делайте новый файл с расширением Java ..."
задание - task
задание1 - task1
задание2 - task2

Имею опыт ООП. Формуировка в том втде что она есть крайне сложна для восприятия и не понятно по сути (Задание сделано практически мащинально не включая головы).
